+++ src/main/java/org/folio/roles/service/loadablerole/LoadableRoleCapabilityAssignmentHelper.java
@@ -93,7 +93,7 @@ public class LoadableRoleCapabilityAssignmentHelper {
     var capabilities = capabilityService.findByPermissionNames(permsByName.keySet());
 
     if (isNotEmpty(capabilities)) {
-      roleCapabilityService.create(roleId, mapItems(capabilities, Capability::getId), false);
+      roleCapabilityService.create(roleId, mapItems(capabilities, Capability::getId), true);
 
       for (Capability cap : capabilities) {
         var perm = permsByName.get(cap.getPermission());
@@ -108,7 +108,7 @@ public class LoadableRoleCapabilityAssignmentHelper {
     var capabilitySets = capabilitySetService.findByPermissionNames(permsByName.keySet());
 
     if (isNotEmpty(capabilitySets)) {
-      roleCapabilitySetService.create(roleId, mapItems(capabilitySets, CapabilitySet::getId), false);
+      roleCapabilitySetService.create(roleId, mapItems(capabilitySets, CapabilitySet::getId), true);
 
       for (CapabilitySet set : capabilitySets) {
         var perm = permsByName.get(set.getPermission());